requirements.txt
asgiref==3.8.1
black==25.1.0
blinker==1.9.0
click==8.1.8
Django==5.1.7
django-filter==25.1
django-leaflet==0.31.0
djangorestframework==3.15.2
environs==14.1.1
Flask==3.1.0
GDAL==3.10.2
geos==0.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.1
MarkupSafe==3.0.2
marshmallow==3.26.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
Werkzeug==3.1.3
